pyproject.toml
[project]
name = "pandasaimongo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy==1.26.4",
    "pandas>=1.5.3",
    "pandasai>=2.3.2",
    "pymongo[srv]==3.13.0",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.5",
]
